+++ src/middleware.ts
@@ -1,4 +1,3 @@
-import { cookies } from 'next/headers'
 import type { NextRequest } from 'next/server'
 import { NextResponse } from 'next/server'
 
@@ -7,8 +6,7 @@ import { ROUTES } from './constants/routes'
 const authRedirectUrls = [ROUTES.PROFILE, ROUTES.ORDERS]
 
 export const middleware = async (request: NextRequest) => {
-   const cookieStore = await cookies()
-   const token = cookieStore.get('authToken')?.value
+   const token = request.cookies.get('authToken')?.value
 
    if (!token && authRedirectUrls.includes(request.nextUrl.pathname)) {
       return NextResponse.redirect(new URL(ROUTES.AUTH, request.url))
